- Bug #2811753: Backslashes weren't correctly written into the user.properties (only one backslash instead of two), thus after restart all backslashes were gone. - Feature Request #2788390: With the user.properties entry "AllowRepositoryModification" one can disable all context menu entries that modify the document repositories. - Feature Request #2812668: Option to save the search history on exit - Sorting by title, filename and path takes numberings into account (e.g. file12 > file2). - Bug: If DocFetcher was installed on the system and crashes, it writes the stacktrace to the application data folder (Windows) or the .docfetcher folder in the home folder (Linux). In the previous version, if DocFetcher ran on Linux, it tried unsuccessfully to write it into /usr/share/docfetcher. - German DocFetcher Manual: "Dateiverwaltung" changed to "Dateimanager"; HTML titles set to "DocFetcher Handbuch..."

+++ DocFetcher/src/net/sourceforge/docfetcher/view/ScopeGroup.java
@@ -20,7 +20,6 @@ import java.util.Set;
 
 import net.sourceforge.docfetcher.Const;
 import net.sourceforge.docfetcher.DocFetcher;
-import net.sourceforge.docfetcher.Event;
 import net.sourceforge.docfetcher.enumeration.Key;
 import net.sourceforge.docfetcher.enumeration.Msg;
 import net.sourceforge.docfetcher.enumeration.Pref;
@@ -30,6 +29,7 @@ import net.sourceforge.docfetcher.model.ResultDocument;
 import net.sourceforge.docfetcher.model.RootScope;
 import net.sourceforge.docfetcher.model.Scope;
 import net.sourceforge.docfetcher.model.ScopeRegistry;
+import net.sourceforge.docfetcher.util.Event;
 import net.sourceforge.docfetcher.util.UtilFile;
 import net.sourceforge.docfetcher.util.UtilGUI;
 import net.sourceforge.docfetcher.util.UtilList;
@@ -104,10 +104,12 @@ public class ScopeGroup extends GroupWrapper {
 		viewerMenu.addSeparator();
 		viewerMenu.addNonEmptyAction(new OpenDirectoryAction(), Key.Enter);
 		viewerMenu.addNonEmptyAction(new ListDocumentsAction(), null);
-		viewerMenu.addSingleElementAction(new CreateSubfolderAction(), Key.ShiftInsert);
-		viewerMenu.addSingleElementAction(new RenameFolderAction(), Key.Rename);
-		viewerMenu.addNonEmptyAction(new DeleteFolderAction(), Key.ShiftDelete);
-		viewerMenu.addSingleElementAction(new PasteIntoFolderAction(), Key.Paste);
+		if (Pref.Bool.AllowRepositoryModification.getValue()) {
+			viewerMenu.addSingleElementAction(new CreateSubfolderAction(), Key.ShiftInsert);
+			viewerMenu.addSingleElementAction(new RenameFolderAction(), Key.Rename);
+			viewerMenu.addNonEmptyAction(new DeleteFolderAction(), Key.ShiftDelete);
+			viewerMenu.addSingleElementAction(new PasteIntoFolderAction(), Key.Paste);
+		}
 		viewerMenu.addSeparator();
 		viewerMenu.addRootAction(new ToggleDeleteOnExitAction(), null);
 		viewerMenu.setManagedActionsEnabled(false);
